osd/PGBackend: stop at seeing the first auth object
in this change, we choose the first good object when looking for an auth
object. before this change, we keep on checking all the collected scrub maps
for the given object even after a good one is found. but we only need a single
auth object here. so it would be better to stop at seeing the first good
object for the authoritative copy.

Signed-off-by: Kefu Chai <kchai@redhat.com>